+++ src/App.jsx
@@ -1,9 +1,25 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
-import Navigation from "./routes/Navigation";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import { ItemListContainer, ItemDetailContainer } from "./components"; // Barrel de componentes
+
+const routes = createBrowserRouter([
+  {
+    path: "/",
+    element: <ItemListContainer greeting="Welcome to Esto no es Moda👠" />,
+  },
+  {
+    path: "/category/:id",
+    element: <ItemListContainer greeting="Products from category" />,
+  },
+  {
+    path: "/item/:id",
+    element: <ItemDetailContainer />,
+  },
+]);
 
 function App() {
-  return <Navigation />;
+  return <RouterProvider router={routes} />;
 }
 
 export default App;
