+++ src/api/products.js
@@ -1,8 +1,10 @@
-export const getProductsByCategory = async (categoryId) => {
+const url = "https://rociocabrera.github.io/api_estonoesmoda/products.json";
+
+export const getProductsByCategoryId = async (categoryId) => {
   try {
-    const response = await fetch("https://rociocabrera.github.io/api_estonoesmoda/products.json");
+    const response = await fetch(url);
     const products = await response.json();
-    return products.filter((product) => product.category.toString() === categoryId);
+    return products.filter((product) => product.category.toString() === categoryId.toString());
   } catch (error) {
     console.log(error);
   }
@@ -10,7 +12,7 @@ export const getProductsByCategory = async (categoryId) => {
 
 export const getProductById = async (id) => {
   try {
-    const response = await fetch("https://rociocabrera.github.io/api_estonoesmoda/products.json");
+    const response = await fetch(url);
     const product = await response.json();
     return product.find((product) => product.id.toString() === id);
   } catch (error) {
@@ -20,7 +22,7 @@ export const getProductById = async (id) => {
 
 export const getAllProducts = async () => {
   try {
-    const response = await fetch("https://rociocabrera.github.io/api_estonoesmoda/products.json");
+    const response = await fetch(url);
     const products = await response.json();
     return products;
   } catch (error) {
@@ -28,4 +30,4 @@ export const getAllProducts = async () => {
   }
 };
 
-export default { getProductsByCategory, getProductById, getAllProducts };
+export default { getProductsByCategoryId, getProductById, getAllProducts };
